Generalize to read filter from file or stdin.
